+++ changelog.md
@@ -2,6 +2,8 @@
 
 ## 0.0.2
 
+* index.js reformat (80 -> 120)
+* description
 * README.md
 
 ## 0.0.1